bug fix
In the case of RIGID Transfo (Similarity and RIGID), in the case of 2D transformation , R (2,2) is computed at -1 . For rhe meantime the fixture is to replace this value by 1 when it is -1. But need to be further investigated.
Note that only affine transformation support Flipping.
Also updated the binary to version 2.0.1