2D 3D consideration
Back to comutation of 3D matrix even for 2D transform. In ec-clem 3D to 2 will produce 2D output (at the slice where points were placed). In ecclemtransform, the same, but if the transform was computed on 2D data (max projection or extracted slice), applying the transfo to a data set with the same dimenion but Z (full slice ) will apply the 2D transform to all slice of the 3D dataset, resulting in a 3D dataset. TODO: ellipse error generation has not been tested yet.